+++ discount_integration_test.go
@@ -33,7 +33,9 @@ func TestDiscounts(t *testing.T) {
 	} else if discount.Name != "test_discount_name" {
 		t.Fatalf("expected Name to be %s, was %s", "test_discount_name", discount.Name)
 	} else if discount.NeverExpires != true {
-		t.Fatalf("expected NeverExpires to be %v, was %v", true, discount.NeverExpires)
+		t.Fatalf("expected NeverExpires to be %t, was %t", true, discount.NeverExpires)
+	} else if discount.CurrentBillingCycle != 0 {
+		t.Fatalf("expected current billing cycle to be %d, was %d", 0, discount.CurrentBillingCycle)
 	} else if discount.Description != "A test discount" {
 		t.Fatalf("expected Description to be %s, was %s", "A test discount", discount.Description)
 	}